Restructure GUI modules, update requirements and READme file
