+++ gui/services/duty_cycle.py
