+++ requirements.txt
@@ -1,24 +1,3 @@
-astropy==6.1.0
-astropy-iers-data==0.2024.5.6.0.29.28
-click==8.1.7
-contourpy==1.2.1
-cycler==0.12.1
-fonttools==4.51.0
-kiwisolver==1.4.5
-matplotlib==3.8.4
-numpy==1.26.4
-packaging==24.0
-pillow==10.3.0
-pyerfa==2.0.1.4
-pyparsing==3.1.2
-PyQt5==5.15.9
-pyqt5-plugins==5.15.9.2.3
-PyQt5-Qt5==5.15.2
-PyQt5-sip==12.12.1
-pyqt5-tools==5.15.9.3.3
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-PyYAML==6.0.1
-qt5-applications==5.15.2.2.3
-qt5-tools==5.15.2.1.3
-six==1.16.0
+numpy>=1.26,<2.0
+astropy>=6.1,<7.0
+PyQt5>=5.15,<6.0
